requirements.txt
anthropic>=0.40.0
google-api-python-client>=2.100.0
PyGithub>=2.1.1
requests>=2.31.0
